+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Initial set of unit tests for `fsictools` module.
+
 ### Changed
 
 - Code generator now removes trailing space after opening brackets; and leading